Add the screenlock after suspend hook to fix the ru-pass issue
I.e. when you leave a russian keyboard layout before the machine
falls asleep, you are locked out when it wakes up,
because there is no way to change the layout in the screen locker.

Fixed by disabling the GDM screen lock, and using the
systemd hook that runs slock (with setxkbmap us before that).